requirements_cpu.txt
numpy==1.21.2
opencv_python==4.5.5.62
Pillow==9.0.1
pytorch_pretrained_bert==0.6.2
scipy==1.8.0
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.3
ftfy
openai-clip